Introduce MutationFactory
